+++ CHANGELOG.md
@@ -1,8 +1,15 @@
 ## Table Of Contents
 
+* [1.0.1](#1.0.1)
 * [1.0.0](#1.0.0)
 * [0.5.0](#0.5.0)
 * [0.1.0](#0.1.0)
+
+## 1.0.0
+
+- Added umbrella to remove xcode warning.
+- Added EDSDebug macro to show let developers choose whether to show non error logs.
+- Removed DATA session wrapping.
 
 ## 1.0.0